finished; very bad; review of _Left in the Dark_ at https://plus.google.com/103530621949492999968/posts/MJ2Sd15Zjh3 : Having finished reading it, I'm very disappointed. It's not fit to shine Jaynes's boot buckle; while Jaynes's theory has serious problems explaining where bicameralism went, worldwide, and other chronological problems, it at least was an elegant theory that fit a great many facts.
Wright and his co-author, on the other hand, is chock-full of
anecdotes (better ping-pong playing and relaxation playing with your
left hand? really?); data seems misrepresented (I did not recognize
his description of the peg data in the sleep deprivation experiment
with the actual graph); dismissal of possibilities as 'implausible'
which are perfectly plausible (the claim that testosterone can't
damage one hemisphere and help the other comes to mind, as does the
claim that there are no plausible evolutionary theories for evolving
handedness - actually, there's a very elegant competition explanation
which anyone who has done a head-to-head sport like ping-pong or
fencing could appreciate); fully general counter-arguments like the
claim that not appreciating the theory is evidence of being damaged as
the theory predicts; excuse-making for the absence of forest fossils
supporting theory theory; blatantly false claims about bipedalism
mixed in with distraction about the aquatic ape theory; claims about
agricultural longevity which are surely the elementary mistake of
assuming that infant mortality-dominated average lifespans say much
about how adults actually lived; invocation of pseudoscience like
photoreading or Carlos Castaneda or 140-year olds in distant Third
World countries (seriously? seriously?!) or Alzheimers and other
diseases being caused by dehydration; blatant confirmation bias* (who,
interested in people who do not sleep, would mention a Vietnamese
woman and *not* mention fatal familial insomnia?); and of course,
vitamin C megadoses must appear in any work by fruitarians, as does
the claim that cooking reduces the nutritional value of food (which
would come as news to anthropologists, who regard the introduction of
fire as possibly the key breakthrough to enabling large expensive
brains - except that right, the authors believe that *fruit* is what
enabled large brains, hm, I wonder why they might not discuss the
caloric value of cooking food... perhaps it's because cooked food
doesn't 'raise micro-electric potentials throughout the body').

* which becomes ever more important in data-rich noisy fields like
psychology; I actually pointed this out for dual n-back recently on
Hacker News: http://news.ycombinator.com/item?id=4204653

The use of autists, idiot savants, and TMS is an interesting topic in
its own right, but inadvertently sabotages the thesis. There's a legal
saying, 'hard cases make bad law'; in neuroscience, damaged patients
make bad theories. The brain is Turing-complete and can exhibit
arbitrarily complex problems: these patients tend to be fairly unique,
and how are you going to build any kind of theory on them? To the
extent we can infer anything, it's that there is a common trait among
those with access to low-level processing: they are *not smart*. Their
abilities seem to come at the expense of all other abilities or
higher-level processing: Kim Peek may have memorized thousands of
books, but what sort of understanding did this man with an IQ of 89
exhibit of them? Very little. The child prodigy artist may be able to
render remarkably detailed artwork, but a camera is more accurate and
just as artistic as I've ever heard him be. This does not bode well
for any exaltation of the right hemisphere...

The crowning piece of nonsense is probably the 'loop' between fruits,
steroids, and DNA. This idea, lovingly illustrated in multiple graphs,
is fractal nonsense: the more you look at each piece, the worse it
gets. It's as if someone told you that car evolution was steered by
gas stations: you see, a human drives the car to a gas station, the
car is filled up with gas, the gas changes the car, changing how it
drives to the next gas station and how it reacts to the next fillup.
This may initially seem plausible until you look at step 3, which
reads basically like the old Far Side strip - 'And then a miracle
happens'. How exactly do steroids rewrite the DNA? A steroid is not a
little computer or nanobot which can go into a sperm or egg - and
notice the quiet eliding of the difference between somatic and
germ-line cells, it's perfectly possible for any genetic or epigenetic
changes to not be passed on - and rewrite the DNA as it wishes. If
variations in circulating levels of some chemical affect DNA, it must
do so via receptors and proteins set up in advance by gradual
evolution to do specific things - eg. think of how alligators or
crocodiles vary the sex ratio of their offspring in response to
ambient temperatures, this is set up in advance because it's useful,
the temperature doesn't 'just cause' the sex ratio to vary in some
magical way because boy alligators are fiery and analogous to hot
weather while female crocodiles are chillier and aloof. (I'm reminded
of the genie of the lamp: "I wish everyone was happy!" Well, OK, but
what does that *actually* mean and how should the genie do it?) They
are, naturally, as sure of this evolutionary loop as one can be in the
complete absence of all evidence.

While we're on DNA, I was deeply amused that they could propose this
system as their primary explanation, and then throughout the rest of
the book dismiss any further adaptation because a few hundred thousand
years was not enough! I really think that we can adapt to eating meat
in a few hundred thousand years, especially when our primate ancestors
and surviving lineages often do eat meat... It was especially funny to
read "It is highly unlikely that the DNA selection process could have
achieved this rapid result (certainly not on the savannah where
hominids would be subjected to hard and stressful environmental
conditions)"; hard and stressful conditions are exactly the kind of
selection pressure that might drive large increases in any organ or
body part, and allometric scaling in general is one of the easier
things for evolution to change if it's fitness-increasing. (Changing
size is a lot easier than developing an organ from scratch, that's for
sure. Size is a tweak.)

These are only my notes for the first 80 pages. After that, I gave up
and read on in a sort of stunned state. At least it does deign to
include some sort of references (although it doesn't link specific
claims to specific papers, seems kind of skimpy, and is very heavy on
books - which aren't very useful in the sense that now you have an
entire book to read just to check a few claims).

It helps that I've read up on many of the same topics for entirely
different reasons, so I have a good idea of the general framework of
various fields and where they're making huge extrapolations or passing
over contrary evidence. 'What is good in it is not original, and what
is original is not good.'

The disturbing thing for me is that so many of the references and
facts are familiar, and it's written fairly well. I'm reminded of
Anatole France:

> What frightens us most in a madman is his sane conversation.