default.html: BLR/AB: reinstate ABalytics's cookie memory for consistency GA continues to report the BLR data by *visit*, not by *page-view*; a visit is a cluster of page-views.
What seems to be happening is that I am randomizing on each individual page load,
but only the randomization of the *first* page-load gets recorded by GA. So
someone who gets tagged 'blue' on the first page will continue to count for
'blue' even if their next page is 'gray6'. This is serious measurement error
which will bias effects towards zero by mixing effects from multiple variants.

As much as I dislike the extra complexity, it seems to be necessary here.